backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.5",
    "python-dotenv>=0.19.0",
    "sqlalchemy>=1.4.23",
    "pydantic>=1.8.2",
    "python-dateutil>=2.8.2",
    "email-validator>=1.1.3",
    "pandas>=2.3.0",
    "psycopg2>=2.9.10",
]
[tool.ruff]
select = ["E", "W", "F", "B"]
ignore = ["E501"]

[tool.mypy]
files = ["main.py", "models.py", "utils/", "routes/"]
ignore_missing_imports = true
